backend/app/services/ai_service.py
"""
AI agent service that handles chat interactions.

TEMPLATE INSTRUCTIONS:
---------------------
This file provides a template for integrating an AI service.
To use with your preferred AI model:
1. Replace the process_message method with your AI integration
2. Update the configure method to set up your AI client
3. Add any additional methods needed for your specific AI service
"""
import asyncio
import logging
from typing import Dict, Any, List

logger = logging.getLogger(__name__)


class AIAgentService:
    """Service for interacting with AI models"""
    
    def __init__(self):
        self.api_key: str = ""
        # You can add other initialization parameters here
        # Example: self.ai_client = None
        
    def configure(self, api_key: str):
        """Configure the AI agent with an API key
        
        You can extend this method to initialize your AI client:
        Example for OpenAI:
            from openai import OpenAI
            self.client = OpenAI(api_key=api_key)
            
        Example for Gemini:
            import google.generativeai as genai
            genai.configure(api_key=api_key)
        """
        self.api_key = api_key
        logger.info("AI agent configured")
    
    async def process_message(self, message: str, context: Dict[str, Any]) -> str:
        """Process a user message and generate a response
        
        REPLACE THIS METHOD with your actual AI model integration.
        Below is a simple mock implementation for demonstration.
        
        Integration Examples:
        
        1. OpenAI:
            response = await self.client.chat.completions.create(
                model="gpt-3.5-turbo",
                messages=[{"role": "user", "content": message}],
                max_tokens=500
            )
            return response.choices[0].message.content
        
        2. Gemini:
            from google.generativeai import GenerativeModel
            model = GenerativeModel('gemini-pro')
            response = await model.generate_content_async(message)
            return response.text
        
        3. Local LLM via API:
            async with httpx.AsyncClient() as client:
                response = await client.post(
                    "http://localhost:11434/api/generate",
                    json={"prompt": message, "model": "llama2"}
                )
                return response.json()["response"]
        """
        # Simulate thinking time
        await asyncio.sleep(1)
        
        # This is a mock implementation - replace with actual AI integration
        if "hello" in message.lower() or "hi" in message.lower():
            return "Hello! I'm your AI assistant. How can I help you today?"
        
        if "files" in message.lower() or "code" in message.lower():
            return "I can help you analyze your code or explore files in the repository. What would you like to know?"
            
        if "github" in message.lower() or "repo" in message.lower():
            repo = context.get("githubRepo", "unknown repository")
            return f"You're working with the {repo} repository. How can I assist with this codebase?"
        
        # Default response
        return "I'm here to help with your code and repository. Please let me know what you'd like to do."
