backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
websockets==12.0
pydantic==2.6.0
httpx==0.27.0
python-dotenv==1.0.0
pytest==8.0.0
PyGithub==2.2.0
